Tests/LibWeb/Text/expected/canvas/export-offscreencanvas.txt

1.Success: {"size":119,"type":"image/png"}
2.Success: {"size":823,"type":"image/jpeg"}
